Split async_io.py and simplify server start/stop.
